ad-categorization-api/requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
pydantic>=2.5.2
pydantic-settings>=2.1.0
google-generativeai>=0.3.0
langchain-google-genai>=0.0.5
langchain>=0.1.0
python-dotenv>=1.0.0